first commit in indirect: got all the figures working.
so all the figures are now configured in such a way that they load up a bare
page, ajax in the figure html, stick it on to our magic div, and then run all
their code. doesn't seem to have broken anything. Migrating back to things that
work over file:// is going to be 1) a monster, and 2) stupid. but you already
knew that.